+++ src/main/java/com/bgm/webapp/repository/CategoryRepository.java
@@ -1,9 +1,11 @@
-package com.bgm.webapp.data.category;
+package com.bgm.webapp.repository;
 
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+
+import com.bgm.webapp.data.category.Category;
 
 /**
  * CategoryRepository
