+++ src/main/java/com/bgm/webapp/security/SecurityConfig.java
@@ -11,7 +11,7 @@ import org.springframework.security.crypto.factory.PasswordEncoderFactories;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.web.SecurityFilterChain;
 
-import com.bgm.webapp.data.user.UserRepository;
+import com.bgm.webapp.repository.UserRepository;
 
 import lombok.RequiredArgsConstructor;
